PDFs prontos

+++ resources/views/viewsProfessores/tabela.blade.php
@@ -20,9 +20,9 @@
                 <td>{{ $prof->sobrenome }}</td>
 
                 <td>
-                    <a href="/professores/info/{{ $prof->id }}">Info</a>
-                    <a href="/professores/editar/{{ $prof->id }}">Editar</a>
-                    <a href="/professores/delete/{{ $prof->id }}">Deletar</a>
+                    <a href="/professores/info/{{ $prof->id }}" class="btn btn-info">Info</a>
+                    <a href="/professores/editar/{{ $prof->id }}" class="btn btn-primary">Editar</a>
+                    <a href="/professores/delete/{{ $prof->id }}" data-confirm='Tem certeza que deseja excluir o item selecionado?' class="btn btn-danger">Deletar</a>
                 </td>
             </tr>
             @endforeach
@@ -67,5 +67,19 @@
         }
     });
 });
+</script>
+
+<script>
+    $(document).ready(function(){
+        $('a[data-confirm]').click(function(ev){
+            var href = $(this).attr('href');
+            if(!$('#confirm-delete').length){
+                $('body').append('<div class="modal fade" id="confirm-delete" aria-labelledby="exampleModalLabel" aria-hidden="true"><div class="modal-dialog"><div class="modal-content"><div class="modal-header"><h5 class="modal-title" id="exampleModalLabel">Você tem certeza?</h5><button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div><div class="modal-body">Tem certeza que deseja apagar os dados do aluno? Não é possível a recuperação dos dados.</div><div class="modal-footer"><button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button><a class="btn btn-danger" id="dataConfirmOk">Deletar</a></div></div></div></div>');
+            };
+            $('#dataConfirmOk').attr('href', href);
+            $('#confirm-delete').modal({shown:true});
+            return false;
+        });
+    });
 </script>
 @endsection
